feat: 서버 영역 custom fetcher 파일명 및 로직 수정
